sc_agents/user_posts_feeds/requirements.txt
firebase_admin
googlemaps
google-genai
